big update

+++ scripts/02_train_models.py
@@ -7,29 +7,38 @@
     -------
     python 02_train_models.py -c pclean6_hemi-L -m unet3d --cuda 2 --lr 0.0025 --momentum 0.9 -s 1 2 3
 
-    python 02_train_models.py -c pclean50A_hemi-L -m unet3d_d00b01_active --cuda 3 --dropout 0 --lr 0.0025 --momentum 0.9 --active --amount 3 --init 4 -r 2
+python 02_train_models.py -c pclean50A_hemi-L -e env_active_learning.json --test pclean12A_hemi-L -m active_test --cuda 4 --dropout 0 --lr 0.0025 --momentum 0.9 --active --amount 3 --init 4 -r 1 --vs 4 --purge --strategy random
 """
 import os.path as op
-from deepsulci.sulci_labeling.capsul.training import SulciDeepTraining, \
-    SulciActiveDeepTraining
+from deepsulci.sulci_labeling.capsul.training import SulciDeepTraining
+from using_deepsulci.sulci_active_deep_training import SulciActiveDeepTraining
+import using_deepsulci.similarity as sim
 import json
 from os import makedirs
 from datetime import datetime
 import argparse
 import shutil
 import numpy as np
+import matplotlib.pyplot as plt
 
 from using_deepsulci.cohort import Cohort
 from utils import Logger
 import sys
 
 
+def plot_sim_mat(mat, save_as):
+    fig = plt.figure(figsize=(8, 8))
+    plt.imshow(mat, interpolation="nearest", aspect="auto")
+    plt.title("Graph Similarity Matrix")
+    fig.savefig(save_as)
+
+
 def train_cohort(cohort, out_dir, fname, dropout=0, lr=0, momentum=0,
                  batch_size=1, translation_file=None, steps=None, cuda=-1,
                  extend=None, active_mode=False, initial_amount=2,
-                 inclusion_amount=1, max_iter=-1,
+                 inclusion_amount=1, max_iter=-1, retrain_from_scratch=False,
                  inclusion_strategy="proba_avg", test_cohort=None,
-                 resampling=.5, detailed_log=False):
+                 voxel_size=None, detailed_log=False):
     proc = SulciActiveDeepTraining() if active_mode else SulciDeepTraining()
 
     if extend:
@@ -51,12 +60,23 @@ def train_cohort(cohort, out_dir, fname, dropout=0, lr=0, momentum=0,
     # Active learning
     if active_mode:
         graphs = cohort.get_graphs()
-        # TODO: verify this
+
         indexes = np.arange(len(graphs)-1)
         np.random.shuffle(indexes)
         init_graphs = []
         for idx in indexes[:initial_amount]:
-            init_graphs.append(graphs.pop(idx))
+            init_graphs.append(graphs[idx])
+        for g in init_graphs:
+            graphs.remove(g)
+
+        if inclusion_strategy == "representativeness":
+            sim_mat_f = op.join(out_dir, "graphs_similarity_matrix.npy")
+            sim_mat = sim.graphs_similarity_matrix(graphs)
+            np.save(sim_mat_f, sim_mat)
+            plot_sim_mat(sim_mat, sim_mat_f)
+            plt.show()
+            proc.similarity_matrix = sim_mat_f
+
         proc.graphs = init_graphs
         proc.available_graphs = graphs
         proc.graphs_notcut = cohort.get_notcut_graphs()
@@ -64,6 +84,7 @@ def train_cohort(cohort, out_dir, fname, dropout=0, lr=0, momentum=0,
         proc.inclusion_strategy = inclusion_strategy
         proc.test_graphs = test_cohort.get_graphs()
         proc.max_iter = max_iter
+        proc.retrain_from_scratch = retrain_from_scratch
     else:
         proc.graphs = cohort.get_graphs()
         proc.graphs_notcut = cohort.get_notcut_graphs()
@@ -75,9 +96,9 @@ def train_cohort(cohort, out_dir, fname, dropout=0, lr=0, momentum=0,
     proc.step_2 = not steps or (steps and 2 in steps)
     proc.step_3 = not steps or (steps and 3 in steps)
     proc.step_4 = not steps or (steps and 4 in steps)
-    #bool(len(cohort.get_notcut_graphs()))
+    # bool(len(cohort.get_notcut_graphs()))
 
-    proc.resampling_ratio = resampling
+    proc.voxel_size = voxel_size
     proc.batch_size = int(batch_size)
     proc.dropout = dropout
     proc.learning_rate = lr
@@ -142,14 +163,20 @@ def main():
                         help="Number of runs to perform")
     parser.add_argument('--purge', dest='purge', const=True, nargs='?',
                         default=False, help="Delete previous results")
-    parser.add_argument('--resampling', dest='resampling', type=float,
-                        default=0.5, help='Resampling factor')
+    parser.add_argument('--retrain', dest='retrain', const=True, nargs='?',
+                        default=False,
+                        help="Retrain from scratch for active learning")
+    parser.add_argument('--pursue', dest='pursue', const=True, nargs='?',
+                        default=False, help="Continue the training")
+    parser.add_argument('--vs', dest='vs', type=float,
+                        default=None, help='Target voxel size')
     parser.add_argument('--detail', dest='detail', const=True, nargs='?',
                         default=False, help="Detailed log")
     args = parser.parse_args()
 
     # Load environment file
-    env_f = args.env if args.env else op.join(op.split(__file__)[0], "env.json")
+    env_f = args.env if args.env else op.join(
+        op.split(__file__)[0], "env.json")
     env = json.load(open(env_f))
 
     cohorts_dir = op.join(env['working_path'], "cohorts")
@@ -171,7 +198,8 @@ def main():
         test_cohort = None
 
     if args.extends:
-        extend = Cohort(from_json=op.join(cohorts_dir, "cohort-" + args.extends + ".json"))
+        extend = Cohort(from_json=op.join(
+            cohorts_dir, "cohort-" + args.extends + ".json"))
         print("Extends with:", extend.name)
     else:
         extend = None
@@ -183,9 +211,13 @@ def main():
             fname = "cohort-" + cohort.name + "_model-" + args.modelname
             mdl_dir = op.join(outdir, fname, 'run-{:02d}'.format(r+1))
 
-            if args.purge:
-                print("Erasing previous results for run-{:02d}".format(r+1))
-                shutil.rmtree(mdl_dir, ignore_errors=True)
+            if op.isdir(mdl_dir):
+                if args.purge:
+                    print(
+                        "Erasing previous results for run-{:02d}".format(r+1))
+                    shutil.rmtree(mdl_dir, ignore_errors=True)
+                elif not args.pursue:
+                    continue
 
             makedirs(mdl_dir, exist_ok=True)
 
@@ -193,8 +225,8 @@ def main():
             train_cohort(cohort, mdl_dir, fname, args.dropout, args.lr,
                          args.momentum, args.batch, env['translation_file'],
                          args.steps, args.cuda, extend, args.active, args.init,
-                         args.amount, args.max_iter, args.strategy, test_cohort,
-                         args.resampling, args.detail)
+                         args.amount, args.max_iter, args.retrain,
+                         args.strategy, test_cohort, args.vs, args.detail)
     return None
